clean up front-end
