+++ front-end/src/App.css
@@ -6,6 +6,12 @@
   height: 40vmin;
   pointer-events: none;
 }
+
+.App-link {
+    color: #000000;
+    text-decoration: none;
+}
+
 .mainContent{
   margin: 10px;
   margin-top: 30px;
