--- front-end/src/Homepage.css
@@ -1,4 +1,0 @@
-.App-link {
-    color: #000000;
-    text-decoration: none;
-  }
