+++ front-end/src/NavBar.js
@@ -10,11 +10,18 @@ const NavBar = () => {
             <Link className = "App-link" to="/">
                 <Navbar.Brand > NYC ROUTE </Navbar.Brand>
             </Link>
-        </Navbar.Collapse>
+        </Navbar.Collapse> 
         <Navbar.Collapse className="justify-content-end">
-            <Link className = "App-link" to="/signup">
-                <Navbar.Brand > Sign Up </Navbar.Brand>
-            </Link>
+          {!localStorage.getItem("token")? (
+              <Link className = "App-link" to="/signup">
+                  <Navbar.Brand > Sign Up </Navbar.Brand>
+              </Link>
+          ) : (
+              <Link className = "App-link" to="/logout">
+                  <Navbar.Brand > Log Out </Navbar.Brand>
+              </Link>
+          )
+          }
         </Navbar.Collapse>
     </Navbar>
   );
