+++ front-end/src/Signup.js
@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from "react"
 import { Link, Redirect } from "react-router-dom";
 import axios from "axios"
+import NavBar from "./NavBar";
 
 const Signup = (props) => {
     const [status, setStatus] = useState({})
@@ -46,34 +47,37 @@ const Signup = (props) => {
 
     if (!status.success)
         return (
-        <div className="mainContent">
-            <h1>Sign Up</h1>
-            <p>{errorMessage}</p>
-            <p>{status.success}</p>
-            <form onSubmit={handleSubmit}>
-                <label>Username: </label>
-                <input type="text" name="username" placeholder="username" />
-                <br />
-                <br />
-                <label>Password: </label>
-                <input type="password" name="password" placeholder="password" />
-                <br />
-                <br />
-                <label>Re-type Password: </label>
-                <input type="password" name="passwordCheck" placeholder="passwordCheck" />
-                <br />
-                <br />
-                <input type="submit" value="Sign Up" />
-            </form>
-            <br />
-            <br />
-            <p>
-                If you already have an account you can
-                <br />
-                <Link  to="/login">
-                Login In
-                </Link>
-            </p>
+        <div>
+          <NavBar />
+          <div className="mainContent">
+              <h1>Sign Up</h1>
+              <p>{errorMessage}</p>
+              <p>{status.success}</p>
+              <form onSubmit={handleSubmit}>
+                  <label>Username: </label>
+                  <input type="text" name="username" placeholder="username" />
+                  <br />
+                  <br />
+                  <label>Password: </label>
+                  <input type="password" name="password" placeholder="password" />
+                  <br />
+                  <br />
+                  <label>Re-type Password: </label>
+                  <input type="password" name="passwordCheck" placeholder="passwordCheck" />
+                  <br />
+                  <br />
+                  <input type="submit" value="Sign Up" />
+              </form>
+              <br />
+              <br />
+              <p>
+                  If you already have an account you can
+                  <br />
+                  <Link  to="/login">
+                  Login In
+                  </Link>
+              </p>
+          </div>
         </div>
         )
     // otherwise, if the user has successfully logged-in, redirect them to a different page
