+++ front-end/src/SubwayLinesInfo.js
@@ -3,7 +3,10 @@ import Button from "react-bootstrap/Button";
 
 import React, { useState, useEffect } from "react";
 import axios from "axios";
+
 import SubwayLinesInfoItem from "./SubwayLinesInfoItem.js";
+import NavBar from "./NavBar";
+
 import "./SubwayLinesInfo.css";
 
 const SubwayLinesInfo = (props) => {
@@ -76,13 +79,16 @@ const SubwayLinesInfo = (props) => {
 
   return (
     <div>
-        <h1> {dataStation.line} Line Info </h1>
-        <Button onClick={reverseOrdering} id = "listOrder"> Reverse </Button>
-        <Accordion id={reverse[1]}>
-          {data.map((item) => (
-            <SubwayLinesInfoItem className="item" key={item.id} details={item} />
-          ))}
-        </Accordion>
+      <NavBar />
+      <div className = "mainContent">
+          <h1> {dataStation.line} Line Info </h1>
+          <Button onClick={reverseOrdering} id = "listOrder"> Reverse </Button>
+          <Accordion id={reverse[1]}>
+            {data.map((item) => (
+              <SubwayLinesInfoItem className="item" key={item.id} details={item} />
+            ))}
+          </Accordion>
+      </div>
     </div>
   );
 };
