+++ front-end/src/SubwayStation.js
@@ -7,6 +7,7 @@ import axios from "axios";
 
 import "./SubwayStation.css";
 import LineCard from "./LineCard";
+import NavBar from "./NavBar";
 
 const EXPRESS_DOMAIN = "http://localhost:4000";
 
@@ -67,23 +68,28 @@ const SubwayStation = (props) => {
 
   if (station.name && station.routes.length === 0) {
     return (
-      <div className="container">
-        <h1 className="stationName">{station.name}</h1>
-        <div className="buttonsWrapper">
-          <Link className="App-link" to="/stations">
-            <Button variant="danger">Back</Button>
-          </Link>
-          <Button
-            variant="primary"
-            onClick={() => {
-              setRefresh(true);
-            }}
-          >
-            Refresh
-          </Button>
-        </div>
-        <div className="no-routes">
-          No trains are arriving within the next hour.
+      <div>
+        <NavBar />
+        <div className="mainContent">
+          <div className="container">
+            <h1 className="stationName">{station.name}</h1>
+            <div className="buttonsWrapper">
+              <Link className="App-link" to="/stations">
+                <Button variant="danger">Back</Button>
+              </Link>
+              <Button
+                variant="primary"
+                onClick={() => {
+                  setRefresh(true);
+                }}
+              >
+                Refresh
+              </Button>
+            </div>
+            <div className="no-routes">
+              No trains are arriving within the next hour.
+            </div>
+          </div>
         </div>
       </div>
     );
@@ -91,64 +97,74 @@ const SubwayStation = (props) => {
 
   if (showFailure) {
     return (
-      <div className="container">
-        <Alert
-          variant="danger"
-          onClose={() => setShowFailure(false)}
-          dismissible
-        >
-          <p className="alertmsg">Error with fetching train data</p>
-        </Alert>
-        <div className="buttonsWrapper">
-          <Link className="App-link" to="/stations">
-            <Button variant="danger">Back</Button>
-          </Link>
-          <Button
-            variant="primary"
-            onClick={() => {
-              setRefresh(true);
-            }}
-          >
-            Refresh
-          </Button>
+      <div>
+        <NavBar />
+        <div className="mainContent">
+          <div className="container">
+            <Alert
+              variant="danger"
+              onClose={() => setShowFailure(false)}
+              dismissible
+            >
+              <p className="alertmsg">Error with fetching train data</p>
+            </Alert>
+            <div className="buttonsWrapper">
+              <Link className="App-link" to="/stations">
+                <Button variant="danger">Back</Button>
+              </Link>
+              <Button
+                variant="primary"
+                onClick={() => {
+                  setRefresh(true);
+                }}
+              >
+                Refresh
+              </Button>
+            </div>
+          </div>
         </div>
       </div>
     );
   }
 
   return (
-    <div className="container">
-      {showSuccess ? (
-        <Alert variant="info" onClose={() => setSuccess(false)} dismissible>
-          <p className="alertmsg">Updated just now</p>
-        </Alert>
-      ) : (
-        ""
-      )}
-      <h1 className="stationName">{station.name}</h1>
-      <div className="buttonsWrapper">
-        <Link className="App-link" to="/stations">
-          <Button variant="danger">Back</Button>
-        </Link>
-        <Button
-          variant="primary"
-          onClick={() => {
-            setRefresh(true);
-          }}
-        >
-          Refresh
-        </Button>
-      </div>
-      <div className="cardsWrapper">
-        {station.routes.map((line) => {
-          return (
-            <LineCard
-              key={line}
-              line={line}
-              traintimes={station.traintimes[line]}
-            />
-          );
-        })}
+    <div>
+      <NavBar />
+      <div className="mainContent">
+        <div className="container">
+          {showSuccess ? (
+            <Alert variant="info" onClose={() => setSuccess(false)} dismissible>
+              <p className="alertmsg">Updated just now</p>
+            </Alert>
+          ) : (
+            ""
+          )}
+          <h1 className="stationName">{station.name}</h1>
+          <div className="buttonsWrapper">
+            <Link className="App-link" to="/stations">
+              <Button variant="danger">Back</Button>
+            </Link>
+            <Button
+              variant="primary"
+              onClick={() => {
+                setRefresh(true);
+              }}
+            >
+              Refresh
+            </Button>
+          </div>
+          <div className="cardsWrapper">
+            {station.routes.map((line) => {
+              return (
+                <LineCard
+                  key={line}
+                  line={line}
+                  traintimes={station.traintimes[line]}
+                />
+              );
+            })}
+          </div>
+        </div>
       </div>
     </div>
   );
