+++ front-end/src/SubwayStations.js
@@ -7,6 +7,8 @@ import axios from "axios";
 import SubwayStationItem from "./SubwayStationItem";
 import "./SubwayStations.css";
 
+import NavBar from "./NavBar";
+
 const EXPRESS_DOMAIN = "http://localhost:4000";
 
 const SubwayStations = () => {
@@ -33,11 +35,14 @@ const SubwayStations = () => {
   if (loading) {
     return (
       <div>
-        <h1>Stations</h1>
-        <div className="spinnerWrapper">
-          <Spinner className="spinner" animation="border" role="status">
-            <span className="visually-hidden">Loading...</span>
-          </Spinner>
+        <NavBar />
+        <div className = "mainContent">
+          <h1>Stations</h1>
+          <div className="spinnerWrapper">
+            <Spinner className="spinner" animation="border" role="status">
+              <span className="visually-hidden">Loading...</span>
+            </Spinner>
+          </div>
         </div>
       </div>
     );
@@ -46,26 +51,32 @@ const SubwayStations = () => {
   if (showFailure) {
     return (
       <div>
-        <h1>Stations</h1>
-        <Alert
-          variant="danger"
-          onClose={() => setShowFailure(false)}
-          dismissible
-        >
-          <p className="alertmsg">Error with fetching train data</p>
-        </Alert>
+        <NavBar />
+        <div className = "mainContent">
+          <h1>Stations</h1>
+          <Alert
+            variant="danger"
+            onClose={() => setShowFailure(false)}
+            dismissible
+          >
+            <p className="alertmsg">Error with fetching train data</p>
+          </Alert>
+        </div>
       </div>
     );
   }
 
   return (
     <div>
-      <h1>Stations</h1>
-      <ListGroup className="stationsWrapper">
-        {stations.map((st) => {
-          return <SubwayStationItem station={st} />;
-        })}
-      </ListGroup>
+      <NavBar />
+      <div className = "mainContent">
+        <h1>Stations</h1>
+        <ListGroup className="stationsWrapper">
+          {stations.map((st) => {
+            return <SubwayStationItem station={st} />;
+          })}
+        </ListGroup>
+      </div>
     </div>
   );
 };
